fix(ha): extract P2/P3 REST endpoints to client_ext, fix ws.rs issues
- Extract ping, components, events, delete_state, get_history_with_flags,
  calendars, template, error_log, check_config to client_ext.rs (736 lines)
- Remove duplicated implementations and tests from client.rs (1199 lines, under CI limit)
- Add client_ext module to lib.rs
- Fix call_command return type in ws.rs (Result<Value> not Result<(id, Value)>)
- Remove duplicate set_supported_features block from ws.rs
- Fix tracing instrument skip list for set_supported_features
- Replace unwrap() with ok_or_else in ws.rs (clippy compliance)
- Make HomeAssistantClient fields pub(crate) for client_ext access
- Update test signature for set_supported_features
- 145 tests pass, clippy clean
